quell braindead warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7964 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ServerMenu.cxx
@@ -557,7 +557,7 @@ void ServerMenu::updateStatus() {
   } else if (serversFound < serverList.size()) {
     std::vector<std::string> args;
     char buffer [80];
-    sprintf(buffer, "%d", serverList.size());
+    sprintf(buffer, "%d", (unsigned int)serverList.size());
     args.push_back(buffer);
     setStatus("Servers found: {1}", &args);
     pageLabel->setString("");
